requirements.txt
# ── Core simulation ──────────────────────────────────────────────────────────
simpy==4.1.1

# ── Machine State Machine ────────────────────────────────────────────────────
transitions==0.9.2

# ── Data & numerics ──────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0

# ── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3

# ── Dashboard ────────────────────────────────────────────────────────────────
streamlit==1.33.0
plotly==5.21.0

# ── REST API ─────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
httpx==0.27.0

# ── Industrial protocols (optional — comment out if not needed) ───────────────
asyncua==1.1.0          # OPC-UA (requires: pip install asyncua)
paho-mqtt==1.6.1        # MQTT

# ── Optimisation ─────────────────────────────────────────────────────────────
pulp==2.8.0             # Linear programming
optuna==3.6.1           # Hyperparameter / meta-optimisation

# ── Console output ───────────────────────────────────────────────────────────
rich==13.7.1
click==8.1.7

# ── Logging ──────────────────────────────────────────────────────────────────
loguru==0.7.2

# ── Data storage (optional) ──────────────────────────────────────────────────
# influxdb-client==1.40.0
# redis==5.0.4
# sqlalchemy==2.0.29
